+++ modules/manifest.go
@@ -7,11 +7,14 @@ import (
 )
 
 type Manifest struct {
+	OrgName       string   `yaml:"org_name"`
 	Name          string   `yaml:"name"`
+	FriendlyName  string   `yaml:"friendly_name"`
 	Description   string   `yaml:"description"`
 	Category      string   `yaml:"category"`
 	Type          string   `yaml:"type"`
 	Layer         string   `yaml:"layer"`
+	IsPublic      bool     `yaml:"is_public"`
 	ProviderTypes []string `yaml:"provider_types"`
 }
 
@@ -19,7 +22,7 @@ func ManifestFromFile(filename string) (*Manifest, error) {
 	file, err := os.Open(filename)
 	if err != nil {
 		if os.IsNotExist(err) {
-			return nil, fmt.Errorf("could not find module name: manifest file %q does not exist", filename)
+			return nil, fmt.Errorf("module manifest file %q does not exist", filename)
 		}
 	}
 	defer file.Close()
@@ -27,7 +30,7 @@ func ManifestFromFile(filename string) (*Manifest, error) {
 	manifest := Manifest{}
 	decoder := yaml.NewDecoder(file)
 	if err := decoder.Decode(&manifest); err != nil {
-		return nil, fmt.Errorf("error decoding manifest: %w", err)
+		return nil, fmt.Errorf("error decoding module manifest: %w", err)
 	}
 	return &manifest, nil
 }